Store per-group/atom energies in mixed precision; real isolated compute
Energy buffers were float, so accumulating in double and storing to float threw
the precision away — exactly where it matters, since alchemical clash/intermediate
states can produce very large energies that fp32 cannot resolve. Make the grid
per-group/per-atom energy buffers and the isolated-GBSA group energy buffers
mixed (double in mixed/double, float in single), both in the kernels and on the
host (allocation + precision-aware readback). The grid total-energy now carries
double in a double context (relE single-vs-double 0 -> 7e-9, previously masked).
getParticleGroupAtomRawEnergies returns vector<double> across all platforms so
the per-atom raw energies used for alchemical u_kln keep full precision.

Also compute the isolated nonbonded diagonal-Hessian and the isolated-GB pairwise
energy in real instead of float (dropping the (float) position-truncating casts);
the |.|^p / fmax exponents are cast to real to avoid the double-mode overload
ambiguity. single/mixed stay bit-identical; double agrees (GBSA/isolated Hessian
rel ~2e-7 vs single).

+++ openmmapi/include/GridForce.h
@@ -1027,7 +1027,7 @@ class OPENMM_EXPORT_GRIDFORCE GridForce : public OpenMM::Force {
      * @param context  the Context to query
      * @return         vector of raw energies, one per atom across all groups (K*N values)
      */
-    std::vector<float> getParticleGroupAtomRawEnergies(OpenMM::Context& context) const;
+    std::vector<double> getParticleGroupAtomRawEnergies(OpenMM::Context& context) const;
 
     /**
      * Compute Hessian (second derivative) blocks for each atom from the grid potential.
